+++ corehq/ex-submodules/pillowtop/processors/form.py
@@ -1,5 +1,12 @@
+from dateutil import parser
+
+from django.http import Http404
+
+from corehq.apps.app_manager.dbaccessors import get_app
+from corehq.apps.users.models import CouchUser
+from corehq.util.quickcache import quickcache
+
 from .interface import PillowProcessor
-from .utils import mark_has_submission, mark_latest_submission
 
 
 class FormSubmissionMetadataTrackerProcessor(PillowProcessor):
@@ -31,3 +38,34 @@ class FormSubmissionMetadataTrackerProcessor(PillowProcessor):
 
         if user_id and domain and received_on:
             mark_latest_submission(domain, user_id, received_on)
+
+
+@quickcache(['domain', 'build_id'], timeout=60 * 60)
+def mark_has_submission(domain, build_id):
+    app = None
+    try:
+        app = get_app(domain, build_id)
+    except Http404:
+        pass
+
+    if app and not app.has_submissions:
+        app.has_submissions = True
+        app.save()
+
+
+def mark_latest_submission(domain, user_id, received_on):
+    user = CouchUser.get_by_user_id(user_id, domain)
+
+    if not user:
+        return
+
+    try:
+        received_on_datetime = parser.parse(received_on)
+    except ValueError:
+        return
+
+    current_last_submission = user.reporting_metadata.last_submission_date
+
+    if current_last_submission is None or current_last_submission < received_on_datetime:
+        user.reporting_metadata.last_submission_date = received_on
+        user.save()
